replicate_deployment/layoutlmv3/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.21.0
cog>=0.9.0
